requirements.txt
openai
pydantic
pyyaml
python-dotenv
pandas 
matplotlib 
jinja2
requests
PyGithub